feat: completa backend con CORS, enums, correcciones y tests unitarios
- Agrega configuracion CORS para frontend Angular (WebConfig)
- Corrige DatosDetalleCliente agregando campo id
- Corrige ClienteService.eliminar para desactivar registro
- Agrega endpoints /enums/partidas y /enums/unidades-medida
- Agrega endpoints DELETE faltantes en NotaVenta y OrdenCompra
- Corrige typo ProdutoController -> ProductoController
- Corrige tipo Integer -> Long en DatosDetalleProducto
- Agrega frontend Angular con Material Design (CRUD completo)
- Crea 66 tests unitarios JUnit5 para controllers, services y helpers
- Corrige test base para usar perfil 'test' con H2

+++ src/main/java/com/laveronica/siscontrol/domain/productos/dto/DatosDetalleProducto.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 public record DatosDetalleProducto(
 
         @NotNull
-        Integer id,
+        Long id,
 
         @NotNull
         String nombre,
@@ -18,7 +18,7 @@ public record DatosDetalleProducto(
         String partida,
 
         @NotNull
-        Integer categoriaId,
+        Long categoriaId,
 
         @NotNull
         String unidadMedida,
@@ -34,10 +34,10 @@ public record DatosDetalleProducto(
 
     public DatosDetalleProducto(Producto producto){
         this(
-                producto.getId().intValue(),
+                producto.getId(),
                 producto.getNombre(),
                 producto.getPartida().name(),
-                producto.getCategoria().getId().intValue(),
+                producto.getCategoria().getId(),
                 producto.getUnidadMedida().name(),
                 producto.getPrecioCompra(),
                 producto.getPrecioVenta()
